15-3-2019 commit

+++ app/Http/Controllers/TaxiController.php
@@ -9,6 +9,9 @@ use Illuminate\Support\Facades\Validator;
 use App\Http\Resources\TaxiResource;
 use Illuminate\Database\Eloquent\ModelNotFoundException;  
 
+use \Datetime;
+use \DateTimeZone;
+
 class TaxiController extends Controller
 {
     /**
@@ -33,15 +36,24 @@ class TaxiController extends Controller
      * Registration for a new taxi after verification
      */
     public function register(Request $request) {
-        $driver = Taxi::create([
+        $validator = Validator::make($request->all(), [
+            'platenumber' => 'max:10|unique:taxis'
+        ]);
+        if($validator->fails()) {
+            return response()->json([
+                'success' => 0,
+                'message' => $validator->errors()], 200);
+        } 
+        $taxi = Taxi::create([
             'platenumber' => $request->platenumber,
             'password' => bcrypt($request->password),
             'occupied' => 0,
-            'owner' => $request->id
+            'owner' => $request->id,
+            'accessToken' => md5(uniqid(rand(), true))
         ]);
         return response()->json([
             'success' => 1,
-            'message' => 'Successfully register for Taxi '.$request->platenumber
+            'message' => $taxi->accessToken
         ]);
     }
 
@@ -55,17 +67,19 @@ class TaxiController extends Controller
         } catch(ModelNotFoundException $e) {
             return response()->json([
                 'success' => 0,
+                'taxi' => NULL,
                 'message' => "Taxi not found"]);
         }
         //If the access token is correct and it is not signed by anyone
-        if($request->accessToken == $taki->accessToken && $taxi->occupied == 0) {
+        if($request->accessToken == $taxi->accessToken && $taxi->occupied == 0) {
             //Update the status of the taxi account
             $currentTime = new DateTime('now');
             $currentTime->setTimezone(new DateTimeZone('Asia/Hong_Kong'));
             $taxi->last_login_time = $currentTime->format('Y-m-d H:i:s');
-            $taxi->occupied = 1;
+            // $taxi->occupied = 1;
             $taxi->driver_id = $request->id;
             $taxi->save();
+            
             // Update the status of the driver
             $driver = Driver::find($request->id);
             $driver->occupied = 1;
@@ -73,15 +87,39 @@ class TaxiController extends Controller
             $driver->save();
             return response()->json([
                 'success' => 1,
-                'message' => "Correct"]);
+                'taxi' => new TaxiResource($taxi),
+                'message' => strval($taxi->id)]);
         } else if($taxi->occupied == 1) {   // The taxi is currently signed in by someone
             return response()->json([
                 'success' => 0,
+                'taxi' => NULL,
                 'message' => "This taxi is already signed up by others"]);
         }
         return response()->json([
             'success' => 0,
+            'taxi' => NULL,
             'message' => "Incorrect password"]);    //Incorrect password
+    }
+
+    public function logout(Request $request) {
+        try {
+            $taxi = Taxi::find($request->taxiID);
+            $driver = Driver::find($request->driverID);
+        } catch(ModelNotFoundException $e) {
+            return response()->json([
+                'success' => 0,
+                'message' => "Taxi not found"]);
+        }
+        $taxi->occupied = 0;
+        $taxi->driver_id = 0;
+        $taxi->save();
+        
+        $driver->occupied = 1;
+        $driver->taxi_id = 0;
+        $driver->save();
+        return response()->json([
+            'success' => 1,
+            'message' => "Success"]);
     }
 
     /**
@@ -109,6 +147,12 @@ class TaxiController extends Controller
             return response()->json([
                 'success' => 0,
                 'message' => "Taxi not found"]);
+        }
+        if($taxi->occupied == 1) {
+            return response()->json([
+                'success' => 0,
+                'message' => $taxi->driver_id->username." is currently taking this taxi"
+            ]);
         }
         if(password_verify($request->password, $taxi->password)) {
             $taxi->delete();
